Refactor types out
I had hard time getting generate-json-schema to work correctly with a
tsx file, but it works fine with a ts file.
